Moving rotors Unfinished 'paid' DLC code TODO: turn RW.cpp to RW.c?
git-svn-id: svn://localhost/vcspc/trunk@39 9f344024-d43e-4b27-bde7-3d027e20192d

+++ VCS PC/Text.h
@@ -53,7 +53,7 @@ private:
 		char		name[8];
 		void*		offset;
 	}				tablEntries[200];
-	unsigned int	size;
+	unsigned short	size;
 };
 
 class CText
